services/api/requirements.txt

fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
SQLAlchemy==2.0.36
psycopg[binary]==3.2.1
bcrypt==4.2.0
PyJWT==2.9.0
requests==2.32.3
beautifulsoup4==4.12.3
pydantic[email]==2.9.2
